Componentize app
