+++ src/components/App.js
@@ -1,0 +1,46 @@
+import Header from './Header';
+import SectionOne from './SectionOne';
+import SectionTwo from './SectionTwo';
+import SectionThree from './SectionThree';
+
+// Import images
+// import b1 from '../images/b1.png';
+// import b2 from '../images/b2.png';
+// import b3 from '../images/b3.png';
+// import b4 from '../images/b4.png';
+// import b5 from '../images/b5.png';
+// import b6 from '../images/b6.png';
+// import circle1 from '../images/circle1.png';
+// import footerLogo from '../images/footer-logo.svg';
+// import hand from '../images/hand.png';
+// import icon1 from '../images/icon1.svg';
+// import icon2 from '../images/icon2.svg';
+// import icon3 from '../images/icon3.svg';
+// import p1 from '../images/p1.png';
+// import p2 from '../images/p2.png';
+// import p3 from '../images/p3.png';
+// import p4 from '../images/p4.png';
+// import p5 from '../images/p5.png';
+// import p6 from '../images/p6.png';
+// import p7 from '../images/p7.png';
+// import p8 from '../images/p8.png';
+// import p9 from '../images/p9.png';
+// import p10 from '../images/p10.png';
+// import p11 from '../images/p11.png';
+// import p12 from '../images/p12.png';
+// import pencil from '../images/pencil.png';
+// import photo from '../images/photo.png';
+// import bucket from '../images/v.png';
+
+export default function App() {
+  return (
+    <>
+      <Header />
+      <main>
+        <SectionOne />
+        <SectionTwo />
+        <SectionThree />
+      </main>
+    </>
+  );
+}
