+++ src/components/ColorPalette.js
@@ -1,0 +1,8 @@
+export const colors = {
+  white: '#ffffff',
+  grayWhite: '#f8f9fa',
+  gray: '#8a94a6',
+  lightBlue: '#e6eafc',
+  mediumBlue: '#546ee8',
+  darkBlue: '#243141'
+};
